+++ src/Creational/Factories/Simple/Car.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codenip\Creational\Factories\Simple;
+
+readonly class Car
+{
+    public function __construct(
+        private string $type,
+    ) {}
+
+    public function type(): string
+    {
+        return $this->type;
+    }
+}