+++ src/Creational/Factories/Simple/CarService.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codenip\Creational\Factories\Simple;
+
+readonly class CarService
+{
+    public function __construct(
+        private SimpleCarFactory $simpleCarFactory,
+    ) {}
+
+    public function createCar(string $type): void
+    {
+        $car = $this->simpleCarFactory->create($type);
+    }
+}